+++ src/main/resources/application.properties
@@ -1,3 +1,7 @@
 spring.application.name=skeleton
 remote-apis.petstore.base-path=${PETSTORE_API_BASE_URL}
 logging.level.wiremock.org=DEBUG
+
+logging.level.org.springframework.web=DEBUG
+logging.level.org.springframework.web.client=DEBUG
+logging.level.org.apache.http=DEBUG
