+++ src/test/java/abc/skeleton/rest_client/controller/PetControllerTest_wiremock_docker_with_mappings.java
@@ -13,7 +13,7 @@ import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class PetControllerWireMockDockerWithMappingsTest {
+class PetControllerTest_wiremock_docker_with_mappings {
 
     @LocalServerPort
     private int port;
