+++ src/test/java/abc/skeleton/rest_client/controller/PetControllerTest_wiremock_testcontainers_with_mappings.java
@@ -1,19 +1,44 @@
 package abc.skeleton.rest_client.controller;
 
+import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.BindMode;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.wiremock.integrations.testcontainers.WireMockContainer;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureWireMock
-public class PetControllerWireMockServerWithMappingsTest {
+@Testcontainers
+class PetControllerTest_wiremock_testcontainers_with_mappings {
+
+    private static final int WIREMOCK_PORT = 8080;
+    @Container
+    private static final WireMockContainer wireMockContainer = new WireMockContainer("wiremock/wiremock:latest")
+            .withFileSystemBind("./src/test/resources/__files", "/home/wiremock/__files", BindMode.READ_WRITE)
+            .withFileSystemBind("./src/test/resources/mappings", "/home/wiremock/mappings", BindMode.READ_WRITE)
+            .withExposedPorts(WIREMOCK_PORT);
     @LocalServerPort
     private int port;
+
+    @BeforeAll
+    static void beforeAll() {
+        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
+    }
+
+    @DynamicPropertySource
+    private static void overrideApiUrl(DynamicPropertyRegistry registry) {
+        registry.add("remote-apis.petstore.base-path", wireMockContainer::getBaseUrl);
+    }
+
 
     @Test
     void should_get_pet_3003() {
